feat: python utils
