+++ py-ft/docs/index.rst
@@ -41,7 +41,11 @@ API Reference
    :show-inheritance:
    :member-order: bysource
 
-
+.. automodule:: utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :member-order: bysource
 
 
 Indices and tables
